# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is a case study assessment prepared for TribeHired",
    "version": "1.0.0",
    "title": "TribeHired Case Study",
    "contact": {
      "name": "Mukhlis Kushairi",
      "email": "mukhliskushairi@gmail.com",
      "url": "https://mukhliskushairi.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["auth"],
        "summary": "Registers a user",
        "description": "Registers a single user into the system",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "required": true,
          "description": "Object that needs to be added into the list of users",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Register"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created"
          },
          "400": {
            "description": "Invalid object structure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/BadRequest"
                }
              }
            }
          },
          "409": {
            "description": "Account already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/AccountExists"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["auth"],
        "summary": "Authenticates a user",
        "description": "Logs a user into the system",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "requestBody": {
          "required": true,
          "description": "Object that will be used to authenticate",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "example": "\"eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImFkbWluQGFkbWluLmNvbSIsImlhdCI6MTU4MDY1ODc0NiwiZXhwIjoxNTgwNjU5NjQ2fQ.vgfrq4Pnfn9GKCm0qPKJt95KZyMztuHYsZ5buae6wvo\""
                }
              }
            }
          },
          "400": {
            "description": "Invalid object structure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/BadRequest"
                }
              }
            }
          },
          "401": {
            "description": "Invalid email or password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/InvalidEmailOrPassword"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": ["user"],
        "summary": "Lists all users",
        "description": "List all registered users in the system",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/UserList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string",
                  "example": "Unauthorized"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  },
  "components": {
    "securitySchemes": {
      "JWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "definitions": {
    "Register": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "John Doe"
        },
        "email": {
          "type": "string",
          "example": "johndoe@mail.com"
        },
        "password": {
          "type": "string",
          "example": "123"
        }
      }
    },
    "Login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string",
          "example": "johndoe@mail.com"
        },
        "password": {
          "type": "string",
          "example": "123"
        }
      }
    },
    "BadRequest": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "\"email\" must be a valid email"
        }
      }
    },
    "AccountExists": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Account already exists"
        }
      }
    },
    "InvalidEmailOrPassword": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "example": "Invalid email or password"
        }
      }
    },
    "UserList": {
      "type": "array",
      "items": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "johndoe@mail.com"
          }
        }
      }
    }
  }
}